Default pages
Add default pages (along with the development workflow for those pages) and implement page loading and caching. The concept of root project (and the required configuration) has been replaced by project assumptions. The project ID for projects without configurations are assumed to be the request hostname. Webpack, plugins, and dependencies, along with React, has been updated. A custom build script for the default pages has been implemented. The idea is to build as close to the project requirement as possible with Webpack, then use the build script to finish off the process. These processes includes replacing stylesheet reference link (<link rel="stylesheet") with embedded stylesheet (<style></style>). The Webpack configuration for this project minifies the stylesheets so page loading time should not be an issue.

- Add build and dev scripts for default page dev work
- Update Webpack and dependencies
- Update Webpack plugins and dependencies
- Update React
- Update Webpack config
- Add pages stylesheet
- Add pages Webpack entry script for loading purposes
- Add default error page
- Add default splash page
- Add custom build script for default pages
- Implement page loader module for page loading and caching
- Implement page cache retrieval in the error handler module
- Implement page loading to cache in the server
- Implement server variable replacements
- Implement custom error page verification
- Remove root project configuration requirement
- Implement project ID assumption for projects without configuration (project ID is assumed to be the request hostname)
- Fix config generation bug in the config loader module
- Add documentation build scripts
- Remove previous iteration of default pages assets under src/public/*