Refactor Docker container to match best practices
Refactor container to multi-stage build. Install upgrades and Node on
the base image. Configure Splash directory permissions on the base
image. Implement two similar image stages, one for /dev/con, one for
production. /dev/con and the production image differs only in the
origin of the Splash installation.

Move maintainer label to final stage and refactor to match best
practices. The maintainer label is following a deprecated format.
Change it.

Remove Supervisor. Splash should run immediately when the container
initiates. Installation is done on the build process. As such, Splash
is the only process that will run on the container. Supervisor is
unnecessary. Remove it.

Remove splash-init.sh. The script executes Splash as a child process
and traps signals to redirect to Splash. Supervisor calls to this
script to launch Splash. It is now unnecessary.