Fix DataLoader bugs
Fix __parseProject(). This function add all projects, regardless of
validity, to __projects map. Scope it to valid projects.

Fix __validateProjectRels(). This function consider a project invalid
when both id and dep are invalid. Change it to consider the project
invalid even if only one of id or dep are invalid.