Refactor Docker build
Add .dockerignore. When the build context is not segmented, it's
important to ignore certain directories to avoid problems.

Move Dockerfile to root directory and add local dev environment to
dev container. Archiving the local directory is an unnecessary step.

Remove package:tar:dev script. Archiving the local directory is an
unnecessary. Additionally, tar is not a cross-platform binary.